e-commerce_backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.0
Pillow==10.4.0

# Payment Gateway Dependencies (install when needed)
# razorpay==1.4.2
# stripe==7.8.0
# requests==2.31.0

# For production
# gunicorn==21.2.0
# psycopg2-binary==2.9.9  # For PostgreSQL
# redis==5.0.1  # For caching and sessions